+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Add basic footer
 - Add basic home with general amiibo request
 - Add basic amiibo item component
+- Add basic amiibo list component